♻️refactor: 디렉토리 구조 변환 & 시맨틱 태그 추가
