+++ src/app/types/commit.ts
@@ -1,0 +1,6 @@
+type CommitInfo = {
+  id: string;
+  message: string;
+};
+
+export type { CommitInfo };
