+++ src/app/ui/report-view/main-area/summary/summary-box.tsx
